% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvoutApp.R
\name{conkApp}
\alias{conkApp}
\title{use shiny to explore univariate and multivariate outlier patterns in selected gene sets}
\usage{
conkApp()
}
\description{
use shiny to explore univariate and multivariate outlier patterns in selected gene sets
}
