% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gbmu133}
\alias{gbmu133}
\title{affy hgu133 expression data for TCGA GBM}
\format{12042 x 528 matrix}
\source{
curatedTCGAData extract with full sample id as colname and hugo symbol as rowname
}
\usage{
gbmu133
}
\description{
affy hgu133 expression data for TCGA GBM
}
\examples{
conkout::gbmu133
}
\keyword{datasets}
