% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tcga_gbm}
\alias{tcga_gbm}
\title{MAF for TCGA GBM samples}
\format{maftools MAF object}
\source{
poisonAlien github TCGAMutations package
}
\usage{
tcga_gbm
}
\description{
MAF for TCGA GBM samples
}
\examples{
conkout::tcga_gbm
}
\keyword{datasets}
